(PC-20410) refacto(libsSearch): move getCategoriesFacetFilters to libs/algolia
