+++ src/libs/search/parseSearchParameters.ts
@@ -3,8 +3,8 @@ import { OffersModuleParameters } from 'features/home/types'
 import { LocationType } from 'features/search/enums'
 import { sortCategories } from 'features/search/helpers/reducer.helpers'
 import { SearchState, SearchView } from 'features/search/types'
+import { getCategoriesFacetFilters } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/getCategoriesFacetFilters'
 import { GeoCoordinates } from 'libs/geolocation'
-import { getCategoriesFacetFilters } from 'libs/search/utils'
 import { buildOfferGenreTypesValues } from 'libs/search/utils/buildOfferGenreTypesValues'
 import { GenreTypeMapping, SubcategoryLabelMapping } from 'libs/subcategories/types'
