--- src/libs/search/utils/index.ts
@@ -1,1 +1,0 @@
-export { getCategoriesFacetFilters } from './getCategoriesFacetFilters'
